# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "Food Delivery API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/internal/order-items/validate": {
            "post": {
                "tags": [
                    "Internal Integration"
                ],
                "summary": "Validate requested order items",
                "description": "Internal endpoint to validate if the restaurant is open and if all requested menu items are available. Combines duplicates inside payload automatically.",
                "operationId": "db3be0a1313bc1d6539f06b711576532",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateOrderItemsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation passed successfully (or normalized successfully)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "restaurant": {
                                            "type": "object"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed (e.g., restaurant closed, item unavailable)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/restaurants/{restaurant}/categories": {
            "get": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Get all categories for a restaurant",
                "operationId": "f97c05724631277eb986ba5e37812b44",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Create a category for a restaurant",
                "operationId": "130628e30755a6df44b460a42d56ad70",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMenuCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu category created successfully"
                    }
                }
            }
        },
        "/restaurants/{restaurant}/categories/all": {
            "get": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Get all categories with cursor pagination (for bulk fetch)",
                "operationId": "d3b2313cbae4de3e8f6172fa35cb8f27",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/categories/{category}": {
            "get": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Get category details",
                "operationId": "b532237741ce1133763f7c88cc51cde5",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Update a category",
                "operationId": "270e5c23c5de600e2ebc73cae95b7f6b",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMenuCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu Categories"
                ],
                "summary": "Delete a category",
                "operationId": "7bcdc6f466fc1ab95d09ae1614a4c7c6",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidateOrderItemsRequest": {
                "required": [
                    "restaurant_id",
                    "items"
                ],
                "properties": {
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "required": [
                                "menu_item_id",
                                "qty"
                            ],
                            "properties": {
                                "menu_item_id": {
                                    "type": "integer",
                                    "example": 101
                                },
                                "qty": {
                                    "type": "integer",
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreMenuCategoryRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Minuman Dingin"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateMenuCategoryRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Minuman Dingin updated"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}